[CL 3356170 by Peter Knepley in UT-Clean branch]

+++ Engine/Source/Runtime/CoreUObject/Public/UObject/ObjectResource.h
@@ -218,14 +218,14 @@ struct FObjectExport : public FObjectResource
 	 * The number of bytes to serialize when saving/loading this export's UObject.
 	 * Serialized
 	 */
-	int32         	SerialSize;
+	int64         	SerialSize;
 
 	/**
 	 * The location (into the FLinker's underlying file reader archive) of the beginning of the
 	 * data for this export's UObject.  Used for verification only.
 	 * Serialized
 	 */
-	int32         	SerialOffset;
+	int64         	SerialOffset;
 
 	/**
 	 * The location (into the FLinker's underlying file reader archive) of the beginning of the
